Refactor: recipient address update logic
Moved updateAddress to another class because of Transactional behavior
