+++ backend/src/main/java/com/tcs/dhv/service/AddressChangeService.java
@@ -67,10 +67,10 @@ public class AddressChangeService {
             parcel.getRecipient().setAddress(savedAddress);
             parcelRepository.save(parcel);
 
-        final var description = String.format("Address changed by %s%s",
-            sender.getEmail(),
-            requestDto.requestReason() != null && !requestDto.requestReason().trim().isEmpty() ? ". Reason: " + requestDto.requestReason() : "");
-        parcelStatusHistoryService.addStatusHistory(parcelId, description);
+            final var description = String.format("Address changed by %s%s",
+                sender.getEmail(),
+                requestDto.requestReason() != null && !requestDto.requestReason().trim().isEmpty() ? ". Reason: " + requestDto.requestReason() : "");
+            parcelStatusHistoryService.addStatusHistory(parcelId, description);
 
             log.info("Address changed successfully for parcel: {} from {} to {}", parcelId, oldAddress.getCity(), savedAddress.getCity());
